requirements.txt
streamlit==1.32.0
requests==2.31.0
matplotlib==3.8.2
stripe==7.11.0
pyyaml==6.0.1
python-dotenv==1.0.1
redis==5.0.1
flask==3.0.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
uwsgi==2.0.24
sqlalchemy